Fix more feedback: Bartosz
- Remove Authorino namespace field from DSCI.
- Move around some code in kserve.go to servicemesh_setup.go
