+++ components/kserve/kserve.go
@@ -4,7 +4,6 @@ package kserve
 import (
 	"context"
 	"fmt"
-	"path"
 	"path/filepath"
 	"strings"
 
@@ -20,7 +19,6 @@ import (
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/cluster"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/deploy"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/feature"
-	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/feature/servicemesh"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/monitoring"
 )
 
@@ -231,38 +229,4 @@ func checkDepedentOps(cli client.Client) *multierror.Error {
 		multiErr = multierror.Append(multiErr, err)
 	}
 	return multiErr
-}
-
-func (k *Kserve) configureServiceMesh(dscispec *dsciv1.DSCInitializationSpec) error {
-	if dscispec.ServiceMesh.ManagementState != operatorv1.Managed || k.GetManagementState() != operatorv1.Managed {
-		return nil
-	}
-
-	serviceMeshInitializer := feature.NewFeaturesInitializer(dscispec, k.defineServiceMeshFeatures(dscispec))
-
-	if err := serviceMeshInitializer.Prepare(); err != nil {
-		return err
-	}
-
-	return serviceMeshInitializer.Apply()
-}
-
-func (k *Kserve) defineServiceMeshFeatures(dscispec *dsciv1.DSCInitializationSpec) feature.DefinedFeatures {
-	return func(s *feature.FeaturesInitializer) error {
-		kserve, err := feature.CreateFeature("configure-kserve-for-external-authz").
-			For(dscispec).
-			Manifests(
-				path.Join(feature.KServeDir),
-			).
-			WithData(servicemesh.ClusterDetails).
-			Load()
-
-		if err != nil {
-			return err
-		}
-
-		s.Features = append(s.Features, kserve)
-
-		return nil
-	}
 }
