+++ pkg/feature/templates/servicemesh/authorino/auth-smm.tmpl
@@ -2,7 +2,7 @@ apiVersion: maistra.io/v1
 kind: ServiceMeshMember
 metadata:
   name: default
-  namespace: {{ .Auth.Namespace }}
+  namespace: {{ .AppNamespace }}
 spec:
   controlPlaneRef:
     namespace: {{ .ControlPlane.Namespace }}
