+++ pkg/feature/templates/servicemesh/authorino/deployment.injection.patch.tmpl
@@ -2,7 +2,7 @@ apiVersion: apps/v1
 kind: Deployment
 metadata:
   name: {{ .AuthorinoConfigs.Name }}
-  namespace: {{ .Auth.Namespace }}
+  namespace: {{ .AppNamespace }}
 spec:
   template:
     metadata:
