+++ pkg/feature/templates/servicemesh/authorino/mesh-authz-ext-provider.patch.tmpl
@@ -9,5 +9,5 @@ spec:
       extensionProviders:
       - name: {{ .AppNamespace }}-odh-auth-provider
         envoyExtAuthzGrpc:
-          service: {{ .AuthorinoConfigs.Name }}-authorino-authorization.{{ .Auth.Namespace }}.svc.cluster.local
+          service: {{ .AuthorinoConfigs.Name }}-authorino-authorization.{{ .AppNamespace }}.svc.cluster.local
           port: 50051
